% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAddressTransactions.R
\name{getAddressTransactions}
\alias{getAddressTransactions}
\title{Querying transactions of an address}
\usage{
getAddressTransactions(address)
}
\arguments{
\item{address}{Address}
}
\value{
Data frame of address transactions
}
\description{
Returns a data frame of address transactions
}
