% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTokenPriceHistoryGrouped.R
\name{getTokenPriceHistoryGrouped}
\alias{getTokenPriceHistoryGrouped}
\title{Querying price history of a token address}
\usage{
getTokenPriceHistoryGrouped(address)
}
\arguments{
\item{address}{Address}
}
\value{
Data frame of price history of a token
}
\description{
Returns a data frame of price history of a token
}
