% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTopTokenHolders.R
\name{getTopTokenHolders}
\alias{getTopTokenHolders}
\title{Querying the top token holders of a token}
\usage{
getTopTokenHolders(address, limit = 10)
}
\arguments{
\item{address}{Token Contract Address}

\item{limit}{Maximum number of token holders (1 - 100)}
}
\value{
Data frame containing address of token holders, percentage of ownership of tokens
}
\description{
Returns the top token holders of a token as a data frame
}
