% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTopTokens.R
\name{getTopTokens}
\alias{getTopTokens}
\title{Querying the top 50 most active tokens}
\usage{
getTopTokens()
}
\value{
A list containing data of the top 50 most active tokens in the past 30 days period
}
\description{
Returns the top 50 of the most active tokens for the last 30 days period
}
