% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTxInfo.R
\name{getTxInfo}
\alias{getTxInfo}
\title{Get transaction info}
\usage{
getTxInfo(transaction_hash)
}
\arguments{
\item{address}{Transaction Hash}
}
\value{
List containing transaction info from a transaction hash
}
\description{
Returns transaction info from a transaction hash
}
