% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTopTokenHolders.R
\name{getTopTokenHolders}
\alias{getTopTokenHolders}
\title{Querying the top token holders of a token}
\usage{
getTopTokenHolders(address, limit = 10)
}
\arguments{
\item{address}{Token Contract Address}

\item{limit}{Maximum number of token holders}
}
\value{
Data frame containing address of token holders, percentage of ownership of tokens
getTopTokenHolders()
}
\description{
This returns the top token holders of a token
}
